added first test for bootstrap_testset and corresponding bugfixes
